backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"

dependencies = [
    "fastapi>=0.120.3",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
